fix build path
